Optimize spinlock
